delete files separately to ensure they are gone

+++ functions/server/Backup-Server.psm1
@@ -77,7 +77,13 @@ try {
   }
 
   # Remove any existing temporary directory
-  Remove-Item -Path "$($Backups.Path)\$Type\$((Get-Item $Backups.Saves).Name)" -Force -Recurse
+  if (Test-Path -Path "$($Backups.Path)\$Type\$((Get-Item $Backups.Saves).Name)" -PathType "Container" -ErrorAction SilentlyContinue) {
+    # Remove the files first
+    Get-ChildItem "$($Backups.Path)\$Type\$((Get-Item $Backups.Saves).Name)" -Recurse | Remove-Item -Force
+
+    # Now remove the folders
+    Remove-Item -Path "$($Backups.Path)\$Type\$((Get-Item $Backups.Saves).Name)" -Force -Recurse
+  }
 
   # Create a temporary directory
   $TempDirectory = New-Item -ItemType Directory -Path "$($Backups.Path)\$Type\$((Get-Item $Backups.Saves).Name)"
@@ -98,6 +104,9 @@ try {
   } else {
     Compress-Archive -Path $TempDirectory -DestinationPath "$($Backups.Path)\$Type\$BackupName.zip" -CompressionLevel 'Fastest'
   }
+
+  # Remove the temporary files
+  Get-ChildItem $TempDirectory -Recurse | Remove-Item -Force
 
   # Remove the temporary directory
   Remove-Item -Path $TempDirectory -Force -Recurse
